refactor: Disable swipe navigation in Timetable.js and add "addAgenda" screen to AppStack.js
This commit disables swipe navigation in Timetable.js by setting the "swipeEnabled" property to false. It also adds a new screen called "addAgenda" to AppStack.js, which allows users to add a new task to the agenda. These changes improve the user experience and enhance the functionality of the app.